pyproject.toml
[tool.poetry]
name = "talkversity_backend"
version = "0.1.0"
description = ""
authors = ["abdalrahmansamara <abdsamara_1998@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.8.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
psycopg2 = "^2.9.1"
psycopg2-binary = "^2.9.1"
django-environ = "^0.6.0"
django-cors-headers = "^3.8.0"

[tool.poetry.dev-dependencies]
Django = "^3.2.7"
black = "^21.8b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
